+++ pyupgw/cli/__init__.py
@@ -39,11 +39,15 @@ from .tui import tui
     help="Password for the user. Alternatively can be loaded from the PYUPGW_PASSWORD environment variable.",
 )
 @click.option(
-    "--logging-config",
-    help="YAML file containing logging config for the application"
+    "--logging-config", help="YAML file containing logging config for the application"
 )
 @click.pass_context
-def cli(ctx: click.Context, username: str | None, password: str | None, logging_config: str | None):
+def cli(
+    ctx: click.Context,
+    username: str | None,
+    password: str | None,
+    logging_config: str | None,
+):
     """Command-line interface for Unisenza Plus
 
     When invoked without subcommand, starts the client in interactive mode.